pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shared_ebay"
version = "0.1.0"
description = "A shared library for eBay OAuth and Browse API"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]

[project.urls]
"Homepage" = "https://github.com/sburl/eBayOauth"
